+++ README.md
@@ -1,2 +1,2 @@
-# lby-fix-engine
-A Fix Engine written by java
+# panda-fix
+A Fix Engine written by java, for study purpose
