+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.lby</groupId>
-    <artifactId>lby-fix-engine</artifactId>
+    <groupId>com.panda</groupId>
+    <artifactId>panda-fix</artifactId>
     <version>1.0.0-SNAPSHOT</version>
 
     <build>
