+++ src/main/java/com/panda/fix/constant/FixSessionType.java
@@ -1,4 +1,4 @@
-package com.lby.fixengine.constant;
+package com.panda.fix.constant;
 
 public enum FixSessionType {
     ACCEPTOR,
