+++ src/main/java/com/panda/fix/session/FixSession.java
@@ -1,6 +1,6 @@
-package com.lby.fixengine.session;
+package com.panda.fix.session;
 
-import com.lby.fixengine.constant.FixSessionType;
+import com.panda.fix.constant.FixSessionType;
 
 import java.util.Objects;
